backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
pymupdf
python-dotenv
python-multipart
